move movement code back into the monobehaviour as it needs to be different for different agents.
