+++ Assets/Digital Painting/Scripts/Agent/AIAgentController.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using UnityEngine;
 using wizardscode.agent.movement;
 using wizardscode.environment;
+using wizardscode.utility;
 
 namespace wizardscode.digitalpainting.agent
 {
@@ -20,7 +21,17 @@ namespace wizardscode.digitalpainting.agent
         private List<Thing> visitedThings = new List<Thing>();
         private Thing _thingOfInterest;
         private float timeLeftLookingAtObject = float.NegativeInfinity;
+        internal float timeToNextWanderPathChange = 0;
         
+        private Transform target; // the current target that we are moving towards
+        private RobotMovementController pathfinding;
+        private Transform wanderTarget;
+
+        new public AIMovementControllerSO MovementController
+        {
+            get { return (AIMovementControllerSO)_movementController; }
+        }
+
         override internal void Awake()
         {
             base.Awake();
@@ -53,14 +64,11 @@ namespace wizardscode.digitalpainting.agent
                 Debug.LogWarning(gameObject.name + " is an AI Agent, but it did not have rigidbody. One has been added automatically so that the agent will not be contained by the '" + DEFAULT_BARRIERS_NAME + "'. Consider adding one.");
             }
 
-            RobotMovementController pathfinding = GetComponent<RobotMovementController>();
+            pathfinding = GetComponent<RobotMovementController>();
             if (pathfinding == null)
             {
                 Debug.LogWarning("No RobotMovementController found on " + gameObject.name + ". One has been added automatically, but consider adding one manually so that it may be optimally configured.");
                 pathfinding = gameObject.AddComponent<RobotMovementController>();
-            } else
-            {
-                ((FlyingAIMovementControllerSO)movementController).pathfinding = pathfinding;
             }
         }
 
@@ -119,22 +127,129 @@ namespace wizardscode.digitalpainting.agent
 
             if (ThingOfInterest != null) // Update POI doesn't always find something
             {
-                movementController.target = ThingOfInterest.AgentViewingTransform;
+                target = ThingOfInterest.AgentViewingTransform;
 
-                if (Vector3.Distance(transform.position, movementController.target.position) > ThingOfInterest.distanceToTriggerViewingCamera)
+                if (Vector3.Distance(transform.position, target.position) > ThingOfInterest.distanceToTriggerViewingCamera)
                 {
-                    movementController.Move(transform);
+                    Move();
                 }
                 else
                 {
-                        ViewPOI();
+                    ViewPOI();
                 }
             }
             else
             {
-                movementController.target = null;
-                movementController.Move(transform);
+                target = null;
+                Move();
             }
+        }
+
+        /// <summary>
+        /// Typically the Move method is called from the Update method of the agent controller.
+        /// It is responsible for making a decision about the agents next move and acting upon
+        /// that decision.
+        /// </summary>
+        private void Move()
+        {
+            if (target != null)
+            {
+                if (!GameObject.ReferenceEquals(pathfinding.Target, target))
+                {
+                    pathfinding.Target = target;
+                    timeToNextWanderPathChange = 0;
+                    return;
+                }
+            }
+            else
+            {
+                UpdateWanderTarget();
+
+                if (Vector3.Distance(transform.position, wanderTarget.position) <= pathfinding.minReachDistance)
+                {
+                    UpdateWanderTarget();
+                }
+            }
+        }
+
+        /// <summary>
+        /// Update the wander target, if it is time to do so.
+        /// A new position for the target is chosen within a cone defined by the
+        /// minAngleOfRandomPathChange and maxAngleOfRandomPathChange. Optionally,
+        /// the cone can extend behind the current agent, which has the effect of 
+        /// turning the agent around.
+        /// </summary>
+        /// <param name="turnAround">Position the target behind the agent. By default this is false.</param>
+        private void UpdateWanderTarget()
+        {
+            if (wanderTarget == null)
+            {
+                wanderTarget = ObjectPool.Instance.GetFromPool().transform;
+            }
+
+            timeToNextWanderPathChange -= Time.deltaTime;
+            if (timeToNextWanderPathChange < 0)
+            {
+                Vector3 position = GetValidWanderPosition(transform, 0);
+
+                if (position == Vector3.zero)
+                {
+                    // Was unable to find a valid position in a few tries so skipping for now, will retry on next frame
+                    Debug.LogWarning("Unable to find a valid wander target");
+                    return;
+                }
+
+                wanderTarget.position = position;
+
+                pathfinding.Target = wanderTarget;
+                timeToNextWanderPathChange = Random.Range(MovementController.minTimeBetweenRandomPathChanges, MovementController.maxTimeBetweenRandomPathChanges);
+            }
+        }
+
+        private Vector3 GetValidWanderPosition(Transform transform, int attemptCount)
+        {
+            int maxAttempts = 6;
+            bool turnAround = false;
+
+            attemptCount++;
+            if (attemptCount > maxAttempts / 2)
+            {
+                turnAround = true;
+            }
+            else if (attemptCount > maxAttempts)
+            {
+                return Vector3.zero;
+            }
+
+            Vector3 position;
+            float minDistance = MovementController.minDistanceOfRandomPathChange;
+            float maxDistance = MovementController.maxDistanceOfRandomPathChange;
+
+            Quaternion randAng;
+            if (!turnAround)
+            {
+                randAng = Quaternion.Euler(0, Random.Range(MovementController.minAngleOfRandomPathChange, MovementController.maxAngleOfRandomPathChange), 0);
+            }
+            else
+            {
+                randAng = Quaternion.Euler(0, Random.Range(180 - MovementController.minAngleOfRandomPathChange, 180 + MovementController.maxAngleOfRandomPathChange), 0);
+                minDistance = maxDistance;
+            }
+            transform.rotation = transform.rotation * randAng;
+            position = transform.position + randAng * Vector3.forward * Random.Range(minDistance, maxDistance);
+
+            // calculate the new height 
+            float terrainHeight = Terrain.activeTerrain.SampleHeight(position);
+            float newY = Mathf.Clamp(position.y, terrainHeight + pathfinding.minFlightHeight, terrainHeight + pathfinding.maxFlightHeight);
+            position.y = newY;
+
+            if (attemptCount <= maxAttempts && !pathfinding.Octree.IsTraversableCell(position))
+            {
+                // Debug.LogWarning("Attempt " + attemptCount + " invalid wander location: " + position);
+                position = GetValidWanderPosition(transform, attemptCount);
+            }
+
+            return position;
         }
 
         /// <summary>
